New test file for utils
